backend/api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
aiomysql==0.2.0
python-dotenv==1.0.1
pydantic==2.6.0
aio-pika==9.4.0
websockets==12.0
cryptography==42.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
